Renamed the components to snake case
