+++ src/components/coinstr/nostr-form.vue
